+++ src/CHANGELOG.md
@@ -7,6 +7,7 @@
   - Fix lag after camera use, refs #106
   - Reduce APK size from 21M to 15M, refs #60
   - Add Android device info to error logs, refs #109
+  - Fix back to home on empty account, fixes #115
 
 ## [v20171029]